+++ game/dota_addons/barebones/scripts/vscripts/gamemode.lua
@@ -23,6 +23,7 @@ require('libraries/animations')
 -- This library can be used for performing "Frankenstein" attachments on units
 require('libraries/attachments')
 
+
 -- These internal libraries set up barebones's events and processes.  Feel free to inspect them/change them if you need to.
 require('internal/gamemode')
 require('internal/events')
@@ -31,7 +32,6 @@ require('internal/events')
 require('settings')
 -- events.lua is where you can specify the actions to be taken when any event occurs and is one of the core barebones files.
 require('events')
-
 
 --[[
   This function should be used to set up Async precache calls at the beginning of the gameplay.