+++ game/dota_addons/barebones/scripts/vscripts/libraries/notifications.lua
@@ -1,11 +1,11 @@
-NOTIFICATIONS_VERSION = "0.88"
+NOTIFICATIONS_VERSION = "0.89"
 
 --[[
   Sample Panorama Notifications Library by BMD
 
   Installation
   -"require" this file inside your code in order to gain access to the Notifications class for sending notifications to players, teams, or all clients.
-  -Additionally, ensure that you have the barebones_hud_base.xml, barebones_hud_base.js, and barebones_hud_base.css files in your panorama content folder.
+  -Additionally, ensure that you have the barebones_notifications.xml, barebones_notifications.js, and barebones_notifications.css files in your panorama content folder.
 
   Usage
   -Notifications can be sent to the Top or Bottom notification panel of an individual player, a whole team, or all clients at once.